Complete Left-Leaning Red-Black Tree Implementation
Implemented recursive LLRB balancing strategy. Benchmarks show a considerable
hit to performance on random insertion and deletion. How much of that is due
to the algorithms I used, and how much due to the quality of my implementation,
is unknown. Empirically, the LLRB implementation shows an improvement in
sequential insertion/deletion time over the unbalanced implementation, and an
improvement over std::Vec for large lists when inserting in reverse order.
Benchmarks to follow.